refactor(ts): Make deep file structures
Separate files by their roles, and merge tests.
