+++ deno.json
@@ -1,15 +1,12 @@
 {
   "$schema": "https://deno.land/x/deno/cli/schemas/config-file.v1.json",
   "fmt": { "exclude": ["LICENSE", ".github/**/*.md"] },
-  "test": { "include": ["src/", "test/"] },
   "tasks": {
     "gen": "deno run --allow-env --allow-read --allow-write ./src/main.ts",
     "test": "deno test --allow-env --allow-read --allow-write --parallel --shuffle",
     "cov": "deno task test --coverage && deno coverage --lcov > coverage.lcov"
   },
   "imports": {
-    "@5ouma/opml-generator/libs": "./src/libs/mod.ts",
-    "@5ouma/opml-generator/types": "./src/types/mod.ts",
     "@libs/xml": "jsr:@libs/xml@5.4.13",
     "@std/assert": "jsr:@std/assert@1.0.2",
     "@std/cli": "jsr:@std/cli@1.0.3",
