+++ src/libs/xml/io.ts
@@ -1,0 +1,16 @@
+import { paramCase } from "@wok/case";
+import { format } from "@std/path";
+import { convert } from "./convert.ts";
+import type { List, Lists } from "../types.ts";
+
+export async function writeXML(feeds: Lists, dir: string): Promise<void> {
+  await Deno.mkdir(dir, { recursive: true });
+  feeds.lists.map(async (list: List) => {
+    const file: string = format({
+      dir: dir,
+      name: paramCase(list.name),
+      ext: ".xml",
+    });
+    await Deno.writeTextFile(file, convert(list));
+  });
+}
