--- src/types/opml.ts
@@ -1,6 +1,0 @@
-export type OPMLOutline = {
-  "@title": string;
-  "@text": string;
-  "@xmlUrl": URL;
-  "@type": "rss";
-};
