--- src/types/sites.ts
@@ -1,4 +1,0 @@
-export type site = {
-  type: string;
-  url: URL;
-};
